requirements.txt
selenium
beautifulsoup4
ics
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
